requirements.txt
python-telegram-bot==13.2
telethon
pyrogram
aiohttp
bleach
emoji
future
google_trans_new
nltk
psutil
fontTools
markdown2
Pillow
chatterbot
chatterbot_corpus
PyYAML
requests
spamwatch
sqlalchemy
wikipedia
bs4
speedtest-cli
coffeehouse
jikanpy
heroku3
zalgo-text
pyrate-limiter
psycopg2-binary
hurry.filesize
alphabet_detector
beautifulsoup4
regex
lxml
TgCrypto
cachetools
telegraph
CurrencyConverter
spongemock
geopy
ffmpeg
